improve exception handlers
