+++ src/Templates/Boilerplate/Bit.Boilerplate/src/Client/Boilerplate.Client.Core/Services/ClientExceptionHandlerBase.cs
@@ -6,13 +6,12 @@ using System.Runtime.CompilerServices;
 
 namespace Boilerplate.Client.Core.Services;
 
-public abstract partial class ExceptionHandlerBase : IExceptionHandler
+public abstract partial class ClientExceptionHandlerBase : SharedExceptionHandler, IExceptionHandler
 {
     [AutoInject] protected Bit.Butil.Console Console = default!;
     [AutoInject] protected ITelemetryContext TelemetryContext = default!;
-    [AutoInject] protected ILogger<ExceptionHandlerBase> Logger = default!;
+    [AutoInject] protected ILogger<ClientExceptionHandlerBase> Logger = default!;
     [AutoInject] protected readonly MessageBoxService MessageBoxService = default!;
-    [AutoInject] protected readonly IStringLocalizer<AppStrings> Localizer = default!;
 
     public void Handle(Exception exception,
         Dictionary<string, object?>? parameters = null,
@@ -35,14 +34,7 @@ public abstract partial class ExceptionHandlerBase : IExceptionHandler
 
         using (var scope = Logger.BeginScope(parameters.ToDictionary(i => i.Key, i => i.Value ?? string.Empty)))
         {
-            var exceptionMessageToLog = exception.Message;
-            var innerException = exception.InnerException;
-
-            while (innerException is not null)
-            {
-                exceptionMessageToLog += $"{Environment.NewLine}{innerException.Message}";
-                innerException = innerException.InnerException;
-            }
+            var exceptionMessageToLog = GetExceptionMessageToLog(exception);
 
             if (exception is KnownException)
             {
@@ -54,8 +46,7 @@ public abstract partial class ExceptionHandlerBase : IExceptionHandler
             }
         }
 
-        string exceptionMessageToShow = (exception as KnownException)?.Message ??
-            (isDevEnv ? exception.ToString() : Localizer[nameof(AppStrings.UnknownException)]);
+        string exceptionMessageToShow = GetExceptionMessageToShow(exception);
 
         MessageBoxService.Show(exceptionMessageToShow, Localizer[nameof(AppStrings.Error)]);
 
@@ -63,30 +54,5 @@ public abstract partial class ExceptionHandlerBase : IExceptionHandler
         {
             Debugger.Break();
         }
-    }
-
-    protected Exception UnWrapException(Exception exception)
-    {
-        if (exception is AggregateException aggregateException)
-        {
-            return aggregateException.Flatten().InnerException ?? aggregateException;
-        }
-        else if (exception is TargetInvocationException)
-        {
-            return exception.InnerException ?? exception;
-        }
-
-        return exception;
-    }
-
-    protected bool IgnoreException(Exception exception)
-    {
-        if (exception is KnownException)
-            return false;
-
-        return exception is TaskCanceledException ||
-            exception is OperationCanceledException ||
-            exception is TimeoutException ||
-            (exception.InnerException is not null && IgnoreException(exception.InnerException));
     }
 }
