+++ src/Templates/Boilerplate/Bit.Boilerplate/src/Client/Boilerplate.Client.Maui/Services/MauiExceptionHandler.cs
@@ -7,7 +7,7 @@ namespace Boilerplate.Client.Maui.Services;
 /// Employing Microsoft.Extensions.Logging implementations (like Sentry.Extensions.Logging) will result in
 /// automatic exception logging due to the logger.LogError method call within the ExceptionHandlerBase class.
 /// </summary>
-public partial class MauiExceptionHandler : ExceptionHandlerBase
+public partial class MauiExceptionHandler : ClientExceptionHandlerBase
 {
     protected override void Handle(Exception exception, Dictionary<string, object> parameters)
     {
