+++ src/Templates/Boilerplate/Bit.Boilerplate/src/Client/Boilerplate.Client.Web/Services/WebExceptionHandler.cs
@@ -1,6 +1,6 @@
 ﻿namespace Boilerplate.Client.Web.Services;
 
-public partial class WebExceptionHandler : ExceptionHandlerBase
+public partial class WebExceptionHandler : ClientExceptionHandlerBase
 {
     protected override void Handle(Exception exception, Dictionary<string, object> parameters)
     {
